Make tape optionally return the counts. Fix emitter to have a uniform radiance. Make counts unsigned integers.

+++ include/mitsuba/render/histogram.h
@@ -104,10 +104,10 @@ public:
     const DynamicBuffer<Float> &data() const { return m_data; }
 
     /// Return the underlying counts for every bin
-    DynamicBuffer<Float> &counts() { return m_counts; }
+    DynamicBuffer<UInt32> &counts() { return m_counts; }
 
     /// Return the underlying counts for every bin (const version)
-    const DynamicBuffer<Float> &counts() const { return m_counts; }
+    const DynamicBuffer<UInt32> &counts() const { return m_counts; }
 
     //! @}
     // =============================================================
@@ -187,7 +187,7 @@ protected:
     ScalarPoint2i m_offset;
     std::vector<float> m_wavelength_bins;
     DynamicBuffer<Float> m_data;
-    DynamicBuffer<Float> m_counts;
+    DynamicBuffer<UInt32> m_counts;
 };
 
 MTS_EXTERN_CLASS_RENDER(Histogram)
